refactor: improve SSG, imports, remove storybook from app and create constants folder
