--- packages/dashboard/.storybook/main.ts
@@ -1,18 +1,0 @@
-import type { StorybookConfig } from "@storybook/nextjs";
-
-const config: StorybookConfig = {
-  stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
-  addons: [
-    "@storybook/addon-onboarding",
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@chromatic-com/storybook",
-    "@storybook/addon-interactions",
-  ],
-  staticDirs: ["../public/"],
-  framework: {
-    name: "@storybook/nextjs",
-    options: {},
-  },
-};
-export default config;
