--- packages/dashboard/.storybook/manager.tsx
@@ -1,9 +1,0 @@
-import { addons } from "@storybook/manager-api";
-import { themes } from "@storybook/theming";
-
-addons.setConfig({
-  theme: {
-    ...themes.dark,
-    brandImage: "/assets/logo.svg",
-  },
-});
