--- packages/dashboard/src/components/pages/AccountHome/AccountHome.constants.ts
@@ -1,12 +1,0 @@
-import { FMenuListItem } from "components";
-
-export const MENU_ITEMS_HOME: FMenuListItem[] = [
-  {
-    label: "Sobre",
-    path: "/sobre",
-  },
-  {
-    label: "Serviços",
-    path: "/servicos",
-  },
-];
