+++ packages/dashboard/src/constants/menuItems.ts
@@ -1,5 +1,16 @@
 import { FMenuListItem } from "components";
 
+export const MENU_ITEMS_HOME: FMenuListItem[] = [
+  {
+    label: "Sobre",
+    path: "/sobre",
+  },
+  {
+    label: "Serviços",
+    path: "/servicos",
+  },
+];
+
 export const MENU_ITEMS_DASHBOARD: FMenuListItem[] = [
   {
     label: "Início",
